+++ frontend/src/process/ProcessManager.js
@@ -1,5 +1,5 @@
 import React, {Component} from "react";
-import ProcessTable from "./ProcessTable";
+import ProcessTable from "./list/ProcessTable";
 
 export default class ProcessManager extends Component {
     render() {
